refactor(reposicao): split AdminReposicaoAumentos (544→119) em hook + componentes
Extrai verbatim de src/pages/AdminReposicaoAumentos.tsx:
- useAumentos: 2 queries (fornecedores + aumentos com agregação de itens),
  memos de agrupamento por mês (grupos/isCollapsed/toggleMes) e handlers de
  upload/extração via Gemini Vision
- AtivosAguardandoBanner, AumentosFiltros, AumentosGrupos, UploadDialog,
  types.ts, config.ts (EMPRESA/FORNECEDOR_DEFAULT/ALL/ESTADOS/
  estadoBadgeClass/formatDate)
- +19 testes (helpers puros + presentacionais)

Comportamento preservado 1:1. tsc baseline=0, eslint=0.
